+++ cogment/actor.py
@@ -32,7 +32,6 @@ class ActorClass:
         observation_space,
         observation_delta,
         observation_delta_apply_fn,
-        feedback_space,
     ):
         self.name = id
         self.config_type = config_type
@@ -40,7 +39,6 @@ class ActorClass:
         self.observation_space = observation_space
         self.observation_delta = observation_delta
         self.observation_delta_apply_fn = observation_delta_apply_fn
-        self.feedback_space = feedback_space
 
 
 class ActorClassList: